backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-multipart==0.0.6
google-generativeai>=0.7.0
pdfminer.six==20221105
python-dotenv==1.0.0
psutil==5.9.8
python-docx>=1.1.0

# LangChain stack
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

# Vector DB
faiss-cpu>=1.7.4
numpy>=1.26.0

# Utilities
tiktoken>=0.6.0